+++ plugin/trino-raptor-legacy/src/test/java/io/trino/plugin/raptor/legacy/BaseRaptorConnectorTest.java
@@ -19,11 +19,11 @@ import com.google.common.collect.SetMultimap;
 import io.trino.Session;
 import io.trino.spi.type.ArrayType;
 import io.trino.testing.BaseConnectorTest;
+import io.trino.testing.Flaky;
 import io.trino.testing.MaterializedResult;
 import io.trino.testing.MaterializedRow;
 import io.trino.testing.TestingConnectorBehavior;
 import io.trino.testing.sql.TestTable;
-import io.trino.testng.services.Flaky;
 import org.intellij.lang.annotations.Language;
 import org.testng.SkipException;
 import org.testng.annotations.Test;
