+++ testing/trino-product-tests/src/main/java/io/trino/tests/product/deltalake/TestDeltaLakeIdentityColumnCompatibility.java
@@ -14,7 +14,7 @@
 package io.trino.tests.product.deltalake;
 
 import io.trino.tempto.assertions.QueryAssert;
-import io.trino.testng.services.Flaky;
+import io.trino.testing.Flaky;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
