+++ testing/trino-product-tests/src/main/java/io/trino/tests/product/hive/TestAvroSchemaUrl.java
@@ -18,7 +18,7 @@ import io.trino.tempto.AfterMethodWithContext;
 import io.trino.tempto.BeforeMethodWithContext;
 import io.trino.tempto.hadoop.hdfs.HdfsClient;
 import io.trino.tempto.query.QueryExecutionException;
-import io.trino.testng.services.Flaky;
+import io.trino.testing.Flaky;
 import org.testng.SkipException;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
