+++ testing/trino-product-tests/src/main/java/io/trino/tests/product/hive/TestCsv.java
@@ -16,7 +16,7 @@ package io.trino.tests.product.hive;
 import io.trino.tempto.ProductTest;
 import io.trino.tempto.assertions.QueryAssert.Row;
 import io.trino.tempto.query.QueryResult;
-import io.trino.testng.services.Flaky;
+import io.trino.testing.Flaky;
 import org.testng.annotations.Test;
 
 import java.util.List;
