+++ testing/trino-testing-services/src/main/java/io/trino/testing/Flaky.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.testng.services;
+package io.trino.testing;
 
 import org.intellij.lang.annotations.Language;
